* cp-tree.def (SUBOBJECT): New tree node. 	* cp-tree.h (CLEANUP_P): New macro. 	(SUBOBJECT_CLEANUP): Likewise. 	(keep_next_level): Add parameter. 	(get_temp_regvar): Don't declare. 	(emit_base_init): Remove parameter. 	(expand_aggr_init): Rename to build_aggr_init. 	(expand_vec_init): Rename to build_vec_init. 	(do_pushlevel): Remove. 	(do_poplevel): Likewise. 	(finish_cleanup): New function. 	(finish_subobject): Likewise. 	(stmts_are_full_exprs_p): New variable. 	* decl.c (keep_next_level): Add parameter. 	(cp_finish_decl): Use build_aggr_init, not 	expand_aggr_init.  Use finish_expr_stmt to expand the code. 	(expand_static_init): Use tree-generating, not RTL-generating, 	functions to handle the initialization. 	(start_function): Remove dead code.  Always have a momentary 	obstack inside the function, even before hitting the first curly 	brace. 	(cplus_expand_expr_stmt): Move calls to 	expand_{start,end}_target_temps into semantics.c. 	(cp_function): Add stmts_are_full_exprs_p. 	(push_cp_function_context): Save it. 	(pop_cp_function_context): Restore it. 	* decl2.c (get_temp_regvar): Move to init.c. 	(do_static_initialization): Use build_{aggr,vec}_init. 	(do_static_destruction): Fix typo in comment. 	* dump.c (dequeue_and_dump): Handle INIT_EXPR. 	* except.c (expand_throw): Use create_temporary_var. 	* expr.c (cplus_expand_expr): Use build_{aggr,vec}_init. 	* init.c (expand_vec_init_try_block): Remove. 	(expand_vec_init_catch_clause): Likewise. 	(get_temp_regvar): New function. 	(begin_init_stmts): Likewise. 	(finish_init_stmts): Likewise. 	(perform_member_init): Use build_{aggr,vec}_init.  Build up tree 	structure here. 	(emit_base_init): Likewise.  Remove unused parameter. 	(expand_virtual_init): Likewise. 	(expand_cleanup_for_base): Use finish_subobject. 	(expand_aggr_vbase_init_1): Simplify. 	(construct_virtual_bases): Use tree-generating functions to build 	up initialization. 	(expand_aggr_init): Likewise.  Rename to build_aggr_init. 	(expand_default_init): Likewise. 	(expand_aggr_init_1): Likewise. 	(expand_vec_init): Rename to build_vec_init. 	* method.c (do_build_copy_constructor): Use tree-generating 	functions.  Don't call clear_last_expr. 	(do_build_assign_ref): Likewise. 	(synthesize_method): Call clear_last_expr here. 	* parse.y (base_init): Don't call clear_last_expr here. 	(nodecls): Likewise. 	* pt.c (tsubst_expr): Handle a TRY_BLOCK with CLEANUP_P set. 	* semantics.c (do_pushlevel): Move to here. 	(do_poplevel): Likewise. 	(stmts_are_full_exprs_p): New variable. 	(finish_expr_stmt): Handle logic for temoprary cleanup here. 	(finish_for_stmt): Use finish_expr_stmt. 	(finish_cleanup): New function. 	(finish_function_try_block): Fix indentation. 	(finish_subobject): New function. 	(setup_vtbl_ptr): Call keep_next_level here. 	(finish_stmt_expr): Handle a block with no scope inside the 	statement-expression. 	(expand_stmt): Handle a TRY_BLOCK with CLEANUP_P set.  Handle 	SUBOBJECT. 	* tree.c (search_tree): Handle INIT_EXPR. 	(mapcar): Likewise. 	* typeck.c (build_modify_expr): Don't build an RTL_EXPR. 	* typeck2.c (store_init_value): Change expand_aggr_init to 	build_aggr_init in comment.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@28878 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/cp/semantics.c
@@ -42,11 +42,26 @@
    much easier since it will be able to make use of these routines.  */
 
 static void expand_stmts PROTO((tree));
+static void do_pushlevel PROTO((void));
+static tree do_poplevel PROTO((void));
 
 /* Non-zero if we should generate RTL for functions that we process.
    When this is zero, we just accumulate tree structure, without
    interacting with the back end.  */
 int expanding_p = 1;
+
+/* Non-zero if we should treat statements as full expressions.  In
+   particular, this variable is no-zero if at the end of a statement
+   we should destroy any temporaries created during that statement.
+   Similarly, if, at the end of a block, we should destroy any local
+   variables in this block.  Normally, this variable is non-zero,
+   since those are the normal semantics of C++.
+
+   However, in order to represent aggregate initialization code as
+   tree structure, we use statement-expressions.  The statements
+   within the statement expression should not result in cleanups being
+   run until the entire enclosing statement is complete.  */
+int stmts_are_full_exprs_p = 1;
 
 /* The type of the last expression-statement we have seen.  This is
    required because the type of a statement-expression is the type of
@@ -91,10 +106,18 @@ finish_expr_stmt (expr)
 	       && lvalue_p (expr))
 	      || TREE_CODE (TREE_TYPE (expr)) == FUNCTION_TYPE)
 	    expr = default_conversion (expr);
-	  cplus_expand_expr_stmt (expr);
-	}
 
-      clear_momentary ();
+	  if (stmts_are_full_exprs_p)
+	    expand_start_target_temps ();
+	    
+	  cplus_expand_expr_stmt (expr);
+
+	  if (stmts_are_full_exprs_p)
+	    {
+	      expand_end_target_temps ();
+	      clear_momentary ();
+	    }
+	}
     }
 
   finish_stmt ();
@@ -455,7 +478,7 @@ finish_for_stmt (expr, for_stmt)
       emit_line_note (input_filename, lineno);
       expand_loop_continue_here ();
       if (expr) 
-	cplus_expand_expr_stmt (expr);
+	finish_expr_stmt (expr);
       expand_end_loop ();
     }
 
@@ -647,11 +670,28 @@ finish_try_block (try_block)
     expand_start_all_catch ();  
 }
 
+/* Finish an implicitly generated try-block, with a cleanup is given
+   by CLEANUP.  */
+
+void
+finish_cleanup (cleanup, try_block)
+     tree cleanup;
+     tree try_block;
+{
+  if (building_stmt_tree ()) 
+    {
+      TRY_HANDLERS (try_block) = copy_to_permanent (cleanup);
+      CLEANUP_P (try_block) = 1;
+    }
+  else
+    expand_eh_region_end (protect_with_terminate (cleanup));
+}
+
 /* Likewise, for a function-try-block.  */
 
 void
 finish_function_try_block (try_block)
-     tree try_block;
+     tree try_block; 
 {
   if (building_stmt_tree ())
     RECHAIN_STMTS_FROM_LAST (try_block, TRY_STMTS (try_block));
@@ -673,9 +713,7 @@ finish_handler_sequence (try_block)
   if (building_stmt_tree ())
     RECHAIN_STMTS_FROM_CHAIN (try_block, TRY_HANDLERS (try_block));
   else
-    {
-      expand_end_all_catch ();
-    }
+    expand_end_all_catch ();
 }
 
 /* Likewise, for a function-try-block.  */
@@ -762,6 +800,12 @@ begin_compound_stmt (has_no_scope)
 
   if (!has_no_scope)
     do_pushlevel ();
+  else
+    /* Normally, we try hard to keep the BLOCK for a
+       statement-expression.  But, if it's a statement-expression with
+       a scopeless block, there's nothing to keep, and we don't want
+       to accidentally keep a block *inside* the scopeless block.  */ 
+    keep_next_level (0);
 
   return r;
 }
@@ -904,6 +948,23 @@ add_decl_stmt (decl)
   add_tree (decl_stmt);
 }
 
+/* We're in a constructor, and have just constructed a a subobject of
+   *THIS.  CLEANUP is code to run if an exception is thrown before the
+   end of the current function is reached.   */
+
+void 
+finish_subobject (cleanup)
+     tree cleanup;
+{
+  if (building_stmt_tree ())
+    {
+      tree r = build_min_nt (SUBOBJECT, cleanup);
+      add_tree (r);
+    }
+  else
+    add_partial_entry (cleanup);
+}
+
 /* Bind a name and initialization to the return value of
    the current function.  */
 
@@ -975,13 +1036,18 @@ setup_vtbl_ptr ()
 		  (CTOR_INITIALIZER,
 		   current_member_init_list, current_base_init_list));
       else
-	emit_base_init (current_class_type, 0);
+	emit_base_init (current_class_type);
     }
+
+  /* Always keep the BLOCK node associated with the outermost pair of
+     curley braces of a function.  These are needed for correct
+     operation of dwarfout.c.  */
+  keep_next_level (1);
 }
 
 /* Begin a new scope.  */
 
-void
+static void
 do_pushlevel ()
 {
   if (!building_stmt_tree ())
@@ -989,22 +1055,24 @@ do_pushlevel ()
       emit_line_note (input_filename, lineno);
       clear_last_expr ();
     }
-  pushlevel (0);
   push_momentary ();
-  if (!building_stmt_tree ())
+  if (stmts_are_full_exprs_p)
+    pushlevel (0);
+  if (!building_stmt_tree () && stmts_are_full_exprs_p)
     expand_start_bindings (0);
-}  
+}
 
 /* Finish a scope.  */
 
-tree
+static tree
 do_poplevel ()
 {
   tree t;
 
-  if (!building_stmt_tree ())
+  if (!building_stmt_tree () && stmts_are_full_exprs_p)
     expand_end_bindings (getdecls (), kept_level_p (), 0);
-  t = poplevel (kept_level_p (), 1, 0);
+  if (stmts_are_full_exprs_p)
+    t = poplevel (kept_level_p (), 1, 0);
   pop_momentary ();
   return t;
 }
@@ -1028,7 +1096,7 @@ finish_parenthesized_expr (expr)
 tree 
 begin_stmt_expr ()
 {
-  keep_next_level ();
+  keep_next_level (1);
   /* If we're building a statement tree, then the upcoming compound
      statement will be chained onto the tree structure, starting at
      last_tree.  We return last_tree so that we can later unhook the
@@ -1055,39 +1123,35 @@ finish_stmt_expr (rtl_expr, expr)
       TREE_SIDE_EFFECTS (rtl_expr) = 1;
     }
 
-  if (TREE_CODE (expr) == BLOCK)
-    {
-      /* Make a BIND_EXPR for the BLOCK already made.  */
-      if (building_stmt_tree ())
-	{
-	  /* If the last thing in the statement-expression was not an
-	     expression-statement, then it has type `void'.  */
-	  if (!last_expr_type)
-	    last_expr_type = void_type_node;
-	  result = build_min (STMT_EXPR, last_expr_type, last_tree);
-	  /* FIXME: Do we need this?  */
-	  TREE_SIDE_EFFECTS (result) = 1;
-	}
-      else
-	result = build (BIND_EXPR, TREE_TYPE (rtl_expr),
-			NULL_TREE, rtl_expr, expr);
-      
-      /* Remove the block from the tree at this point.  It gets put
-	 back at the proper place when the STMT_EXPR or BIND_EXPR is
-	 expanded.  */
-      delete_block (expr);
-    }
-  else
-    result = expr;
-
   if (building_stmt_tree ())
     {
+      /* If the last thing in the statement-expression was not an
+	 expression-statement, then it has type `void'.  */
+      if (!last_expr_type)
+	last_expr_type = void_type_node;
+      result = build_min (STMT_EXPR, last_expr_type, last_tree);
+      TREE_SIDE_EFFECTS (result) = 1;
+      
       /* Remove the compound statement from the tree structure; it is
 	 now saved in the STMT_EXPR.  */
       last_tree = rtl_expr;
       TREE_CHAIN (last_tree) = NULL_TREE;
     }
-  
+  else if (expr && TREE_CODE (expr) == BLOCK)
+    {
+      result = build (BIND_EXPR, TREE_TYPE (rtl_expr),
+		      NULL_TREE, rtl_expr, expr);
+      delete_block (expr);
+    }
+  else
+    result = rtl_expr;
+
+  if (expr && TREE_CODE (expr) == BLOCK)
+    /* Remove the block from the tree at this point.  It gets put back
+       at the proper place when the STMT_EXPR or BIND_EXPR is
+       expanded.  */
+    delete_block (expr);
+
   return result;
 }
 
@@ -2082,11 +2146,20 @@ expand_stmt (t)
 
     case TRY_BLOCK:
       lineno = STMT_LINENO (t);
-      begin_try_block ();
-      expand_stmt (TRY_STMTS (t));
-      finish_try_block (NULL_TREE);
-      expand_stmts (TRY_HANDLERS (t));
-      finish_handler_sequence (NULL_TREE);
+      if (CLEANUP_P (t))
+	{
+	  expand_eh_region_start ();
+	  expand_stmt (TRY_STMTS (t));
+	  finish_cleanup (TRY_HANDLERS (t), NULL_TREE);
+	}
+      else
+	{
+	  begin_try_block ();
+	  expand_stmt (TRY_STMTS (t));
+	  finish_try_block (NULL_TREE);
+	  expand_stmts (TRY_HANDLERS (t));
+	  finish_handler_sequence (NULL_TREE);
+	}
       break;
 
     case HANDLER:
@@ -2103,6 +2176,11 @@ expand_stmt (t)
       finish_handler_parms (NULL_TREE);
       expand_stmt (HANDLER_BODY (t));
       finish_handler (NULL_TREE);
+      break;
+
+    case SUBOBJECT:
+      lineno = STMT_LINENO (t);
+      finish_subobject (SUBOBJECT_CLEANUP (t));
       break;
 
     default:
@@ -2167,11 +2245,6 @@ expand_body (fn)
   /* If this is a constructor, we need to initialize our members and
      base-classes.  */
   setup_vtbl_ptr ();
-
-  /* Always keep the BLOCK node associated with the outermost pair of
-     curly braces of a function.  These are needed for correct
-     operation of dwarfout.c.  */
-  keep_next_level ();
 
   /* Expand the body.  */
   expand_stmt (t);
